Represent measurements as structs instead of classes.
My benchmarks have shown that implementing measurements as classes has a
non-negligible cost.  In the worst case, it increases the cost of
monitoring a form by a factor of three.  The idea is that monitoring can be
used also in production, so I decided to go for the fast, less flexible
approach.